refactoring deserialization to remove special-case behavior for "top level" deserializer and "subclass" deserializer (AssignmentDeserializer and the two subclasses, respectively)
The problem with the current setup is that "deserializeAttributes" is called only on the "top level" object, and takes in an already-created assignment object. Both of the subclasses implement a different method that both creates the object and then parses their specific fields. Since these classes are hardcoded to create new instances within their "deserializeAssignmentSubclass" method, they would need to be refactored if a subclass is to be extended off of either of these concrete base classes.

It now works like this: 

-The AssignmentDeserializerFactory returns a subtype of AssignmentDeserializer
-The AssignmentDeserializerFactory then calls .deserializeAttributes(JsonNode node) on the parser, and returns the Assignment object that is returned.

As for the deserializers, it's now possible to have an arbitrary depth of nested Deserializers, instead of 2. Each subclassing deserializer must implement two methods: 

- deserializeAttributes(JsonNode node): this is where a JsonNode node will be passed in. The class needs to check for any values that it is concerned with and populate it into the assignment object. This is where the magic happens - calling getAssignment() within this class will return an object guaranteed to be the type of object this deserializer is concerned with (either the specific class itself or a subclass of it).

In other words, if you have anything to check, the first line of your deserializeAttributes method should be getAssignment(). The returned assignment should be populated with any relevant JSON values.

You must call super.deserializeAttributes() if you want the parent class attributes to be applied. You can call super.deserializeAttributes either before or after you parse the attributes specific to the class.